+++ CHANGELOG.md
@@ -12,19 +12,33 @@ This release accompanies the release of our paper "Global continuation as a comp
   additional quantities after sampling the state space.
 - `continuation_series` now allows for mixed type outputs, so that the fill value
   can be of different type (e.g., `missing`) than the continuation quantity.
+- New API for instructing `global_continuation` on how to sample initial conditions
+  structured around a new abstract type `InitialConditionSampler`.
+  - Concrete types include `RandomICSampler, PrescribedICs, PerParameterICs` and more
+    to come in the future.
 
-## Prior deprecations removed
+## Breaking changes and Deprecations
 
+- **BREAKING**: The `basins_fractions` output structure has been reworked. Now `basins_fractions` always returns only one output regardless of inputs: the fractions.
+  A new function `basins_fractions_labels` returns also the labels.
 - **BREAKING**: All deprecations existing before v2 release have been removed. Add v1.39 explicitly to resolve this if need be.
+- **DEPRECATED**: calling `global_continuation` with `prange, pidx` is deprecated.
+  Use the version where a single `pcurve` is provided instead, by making
+  `pcurve = [Dict(pidx => p) for p in prange]`.
+- **DEPRECATED**: giving a vector of initial conditions, or a sampling function,
+  as the last argument to `global_continuation` is deprecated. Use an instance of a subtype
+  of `InitialConditionSampler` instead.
+- **DEPRECATED**: Passing a vector of initial conditions or a sampling function to
+  `global_continuation` is now deprecated. Use the `InitialConditionSampler` interface.
 
-## Renaming of stability measures
+### Renaming of stability measures
 
 - The concept of "stability measures" has been renamed into "stability quantifiers" throughout the docs
 - The following renames of API are deprecated:
   - `stability_measures_along_continuation` -> `stability_quantifiers_along_continuation`
   - `StabilityMeasuresAccumulator` -> `StabilityQuantifiersAccumulator`
 
-## Renaming of attractor mappers
+### Renaming of attractor mappers
 
 The `AttractorMapper` constructions have been fully renamed. All following renames are deprecated: